+++ src/main/java/com/hrms/model/SubModule1.java
@@ -1,56 +1,17 @@
 package com.hrms.model;
 
-import java.io.Serializable;
-import java.util.Date;
-import java.util.List;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
 
-@Entity
-@Table(name = "M_SUB_MODULE")
-public class SubModule1 implements Serializable {
+public class SubModule1  {
 
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = -6851090654306622861L;
-
-	@Id
-	@Column(name = "SUB_MODULE_CODE")
 	private String subModuleCode;
-
-	@Column(name = "SUB_MODULE_NAME")
+	
 	private String subModuleName;
+	
+	private String moduleCode;
 
-	@ManyToOne
-	@JoinColumn(name = "MODULE_CODE", nullable = false)
-	private Module moduleCode;
-
-	@Column(name = "INS_BY")
-	private String insertedBySubModule;
-
-	@Column(name = "INS_DATE")
-	private Date insertedDateSubModule;
-
-	@Column(name = "UPDATE_BY")
-	private String updateBySubModule;
-
-	@Column(name = "UPDATE_DATE")
-	private Date updatedDateSubModule;
-
-	@Column(name = "ACTIVE_YN")
 	private String acitveSubModule;
-
-	@Column(name = "SEQ_NO")
+	
 	private int seqNoSubModule;
-
-	@OneToMany(mappedBy = "subModuleCode")
-	private List<Program> subModulePrograms;
 
 	public String getSubModuleCode() {
 		return subModuleCode;
@@ -68,52 +29,12 @@ public class SubModule1 implements Serializable {
 		this.subModuleName = subModuleName;
 	}
 
-	public Module getModuleCode() {
+	public String getModuleCode() {
 		return moduleCode;
 	}
 
-	public void setModuleCode(Module moduleCode) {
+	public void setModuleCode(String moduleCode) {
 		this.moduleCode = moduleCode;
-	}
-
-	public String getInsertedBySubModule() {
-		return insertedBySubModule;
-	}
-
-	public void setInsertedBySubModule(String insertedBySubModule) {
-		this.insertedBySubModule = insertedBySubModule;
-	}
-
-	public Date getInsertedDateSubModule() {
-		return insertedDateSubModule;
-	}
-
-	public void setInsertedDateSubModule(Date insertedDateSubModule) {
-		this.insertedDateSubModule = insertedDateSubModule;
-	}
-
-	public String getUpdateBySubModule() {
-		return updateBySubModule;
-	}
-
-	public void setUpdateBySubModule(String updateBySubModule) {
-		this.updateBySubModule = updateBySubModule;
-	}
-
-	public Date getUpdatedDateSubModule() {
-		return updatedDateSubModule;
-	}
-
-	public void setUpdatedDateSubModule(Date updatedDateSubModule) {
-		this.updatedDateSubModule = updatedDateSubModule;
-	}
-
-	public String getAcitveSubModule() {
-		return acitveSubModule;
-	}
-
-	public void setAcitveSubModule(String acitveSubModule) {
-		this.acitveSubModule = acitveSubModule;
 	}
 
 	public int getSeqNoSubModule() {
@@ -124,12 +45,16 @@ public class SubModule1 implements Serializable {
 		this.seqNoSubModule = seqNoSubModule;
 	}
 
-	public List<Program> getSubModulePrograms() {
-		return subModulePrograms;
+	public String getAcitveSubModule() {
+		return acitveSubModule;
 	}
 
-	public void setSubModulePrograms(List<Program> subModulePrograms) {
-		this.subModulePrograms = subModulePrograms;
+	public void setAcitveSubModule(String acitveSubModule) {
+		this.acitveSubModule = acitveSubModule;
 	}
+	
+	
+
+	
 
 }
