+++ src/main/resources/templates/editSubModule.html
@@ -65,7 +65,7 @@ body {
 											<label style="color: red">*</label>
 											<select class="form-control" name="moduleCode"id="moduleCode"  title="Module Name" required aria-required="true">
 												    <option value="">Select Module</option>
-													<option   th:each="module: ${modulesList}" th:value="${module.getModuleCode()}" th:text="${module.getModuleName()}" th:field="${subModuleEdit.moduleCode}"></option>
+													<option   th:each="module: ${modulesList}" th:value="${module.getModuleCode()}" th:text="${module.getModuleName()}" th:field="${subModuleEdit.moduleCode.moduleCode}"></option>
 											     </select>
 										   </div>
